Rename and convert templates to normal yml
During integration it became apparent that the approach to merge
different template files containing redundant information to one recipe
was not feasible. The approach now is to have saparate .yml files
containing the `diagnostics` part of a regular recipe. Preprocessors
needed can go to the `general.yml` file containing also the general
information part of a regular recipe.

Based on this .yml files a single quicklook recipe file is created