Add alternative multi-cube statistics engine
This function operates directly on iris cubes. It supports
lazy evaluation, and requires less error-prone array handling.
However, it requires the input cubes to be very homogeneous, so
for backwards compatibility I kept the old function as well.

This way we can easily test and improve the new function until
we're happy, and then deprecate the old function.